% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_bagging_prim.R
\name{bagging.prim}
\alias{bagging.prim}
\title{PRIM with bagging (bumping)}
\usage{
bagging.prim(
  dtrain,
  dtest,
  deval = dtrain,
  box,
  minpts = 20,
  peel.alpha = 0.05,
  features = "some",
  iter = 30,
  denom = 5
)
}
\arguments{
\item{dtrain}{list, containing training data. The first element contains matrix/data frame of real attribute values.
the second element contains vector of labels 0/1.}

\item{dtest}{list, containing test data. Structured in the same way as \code{dtrain}}

\item{box}{hyperbox, covering data}

\item{minpts}{integer. Minimal number of points in the box for PRIM to continue peeling}

\item{features}{character or integer. Refers to the number of features used in each iteration.
If integer, defines this number.
If "some", calculated as \code{ceiling(sqrt(ncol(dtrain[[1]])))},
if "all", all features are used. Default is "some"}

\item{iter}{integer. The number of iterations (sequences of boxes learned). Default is 50}
}
\value{
list.
\itemize{
\item \code{pr.test} matrix with coverage (recall) in the first column and
density (precision) in the second column, evaluated on \code{dtest}
\item \code{pr.train} matrix with coverage (recall) in the first column and
density (precision) in the second column, evaluated on \code{deval}
\item \code{boxes} list of matrices defining boxes constituting peeling trajectory
}
}
\description{
The function implements PRIM with bagging (bumping)
}
\references{
J. H. Kwakkel and S. C. Cunningham, "Improving scenario discovery by bagging random boxes,"
Technological Forecasting and Social Change, vol. 111, pp. 124-134, 2016.
}
\seealso{
\code{\link{norm.prim}},
\code{\link{rf.prim}}
}
\keyword{models,}
\keyword{multivariate}
