% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_best_interval.R
\name{best.interval}
\alias{best.interval}
\title{BestInterval subgroup discovery algorithm}
\usage{
best.interval(
  dtrain,
  dtest = NULL,
  box,
  depth = 5,
  beam.size = 1,
  keep = 10,
  denom = 6
)
}
\arguments{
\item{dtrain}{list, containing training data. The first element contains matrix/data frame of real attribute values.
the second element contains vector of labels 0/1.}

\item{dtest}{list, containing test data. Structured in the same way as \code{dtrain}. If NULL, the
WRAcc on test data is not computed}

\item{box}{matrix of real. Initial hyperbox, covering data}

\item{depth}{integer or "cv". The maximal number of dimensions restricted.
If "cv", \code{depth} is chosen automatically with 5-fild cross-validation
from a set defined by \code{ncol(dtrain[[1]])} and \code{denom} parameter}

\item{beam.size}{integer.The size of the beam in the beam search}

\item{keep}{integer. If greater than \code{beam.size}, specifies the maximum
number of boxes to be refined at each iteration in case all have equal WRAcc}

\item{denom}{the maximal length of the set of \code{features} values to choose from}
}
\value{
list.
\itemize{
\item \code{qtest} WRAcc value computed from \code{dtest}
\item \code{qtrain} WRAcc value computed from \code{dtrain}
\item \code{box} the hyperbox, with highest value of WRAcc on \code{dtrain}
}
}
\description{
The function applies BestInterval to train data and evaluates its quality on test data
}
\examples{

dtrain <- dtest <- list()
dtest[[1]] <- dsgc_sym[1001:10000, 1:12]
dtest[[2]] <- dsgc_sym[1001:10000, 13]
dtrain[[1]] <- dsgc_sym[1:1000, 1:12]
dtrain[[2]] <- dsgc_sym[1:1000, 13]
box <- matrix(c(0.5,0.5,0.5,0.5,1,1,1,1,0.05,0.05,0.05,0.05,
5,5,5,5,4,4,4,4,1,1,1,1), nrow = 2, byrow = TRUE)

norm.prim.w(dtrain = dtrain, dtest = dtest, box = box, peel.alpha = 0.05)
norm.prim.w(dtrain = dtrain, dtest = dtest, box = box, peel.alpha = c(0.01, 0.03, 0.05, 0.07, 0.09, 0.11, 0.13, 0.15, 0.17, 0.19))
best.interval(dtrain = dtrain, dtest = dtest, box = box, depth = 12, beam.size = 1)
best.interval(dtrain = dtrain, dtest = dtest, box = box, depth = "cv", beam.size = 1)


dx <- bi_test[, 1:4]
dy <- bi_test[, 5]
box = matrix(c(0,0,0,0,1,2,1,1), nrow = 2, byrow = TRUE)
sort(dx[, 3], decreasing = TRUE)[5:6]

best.interval(list(dx, dy), box = box, depth = 3, beam.size = 1)$box
best.interval(list(dx, dy), box = box, depth = 3, beam.size = 4)$box               # clear difference in the fourth dimension

best.interval(list(dx, dy), box = box, depth = 4, beam.size = 4)$box
norm.prim.w(list(dx, dy), box = box, peel.alpha = 0.05)$box
norm.prim.w(list(dx, dy), box = box, peel.alpha = 0.1)$box
}
\references{
Mampaey, Michael, et al. "Efficient algorithms for finding richer subgroup descriptions
in numeric and nominal data." 2012 IEEE 12th International Conference on Data Mining. IEEE, 2012.
}
\seealso{
\code{\link{norm.prim.w}}
}
\keyword{models,}
\keyword{multivariate}
