% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bi_test}
\alias{bi_test}
\title{Dataset for testing BestInterval algorithm}
\format{
bi_test is a matrix with 5 columns and 1000 rows. First 4 columns are attributes; the last column - label
}
\usage{
data(bi_test)
}
\description{
Dataset for testing BestInterval algorithm
}
\keyword{datasets}
