% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_consistencyd.R
\name{consistency.d}
\alias{consistency.d}
\title{Consistency based on data}
\usage{
consistency.d(box1, box2, ds)
}
\arguments{
\item{box1}{real. Matrix defining the first box to compare}

\item{box2}{real. Matrix defining the second box to compare}

\item{ds}{data to be used for calculations. Matrix/data frame of real attribute values}
}
\value{
a number, quantidying consistency between two boxes based on data
}
\description{
The function calculates consistency for two N-dimensional hyperboxes based on data
}
\examples{

dtrain <- dtest <- list()
dtest[[1]] <- dsgc_sym[5001:10000, 1:12]
dtest[[2]] <- dsgc_sym[5001:10000, 13]
dtrain[[1]] <- dsgc_sym[1:5000, 1:12]
dtrain[[2]] <- dsgc_sym[1:5000, 13]
box <- matrix(c(0.5,0.5,0.5,0.5,1,1,1,1,0.05,0.05,0.05,0.05,
5,5,5,5,4,4,4,4,1,1,1,1), nrow = 2, byrow = TRUE)

set.seed(1)
res1 <- norm.prim(dtrain = dtrain, dtest = dtest, box = box)
res2 <- norm.prim(dtrain = dtest, dtest = dtrain, box = box)
box1 <- res1[[3]][[length(res1[[3]])]]
box2 <- res2[[3]][[length(res2[[3]])]]
consistency.d(box1, box2, dtrain)
consistency.d(box1, box2, dtest)
}
\seealso{
\code{\link{consistency.v}},
\code{\link{consistency.l}}
}
