% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_consistencyl.R
\name{consistency.l}
\alias{consistency.l}
\title{Linear consistency}
\usage{
consistency.l(box1, box2, box.init)
}
\arguments{
\item{box1}{real. Matrix defining the first box to compare}

\item{box2}{real. Matrix defining the second box to compare}

\item{box.init}{real. Matrix describing the initial box containing all data}
}
\value{
a number, quantidying linear consistency between boxes
}
\description{
The function calculates linear consistency for two N-dimensional hyperboxes
}
\examples{

dtrain <- dtest <- list()
dtest[[1]] <- dsgc_sym[5001:10000, 1:12]
dtest[[2]] <- dsgc_sym[5001:10000, 13]
dtrain[[1]] <- dsgc_sym[1:5000, 1:12]
dtrain[[2]] <- dsgc_sym[1:5000, 13]
box <- matrix(c(0.5,0.5,0.5,0.5,1,1,1,1,0.05,0.05,0.05,0.05,
5,5,5,5,4,4,4,4,1,1,1,1), nrow = 2, byrow = TRUE)

set.seed(1)
res1 <- norm.prim(dtrain = dtrain, dtest = dtest, box = box)
res2 <- norm.prim(dtrain = dtest, dtest = dtrain, box = box)
box1 <- res1[[3]][[length(res1[[3]])]]
box2 <- res2[[3]][[length(res2[[3]])]]
consistency.l(box1, box2, box)
}
\seealso{
\code{\link{consistency.v}},
\code{\link{consistency.d}}
}
