% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_get_dominant.R
\name{get.dominant}
\alias{get.dominant}
\title{dominating boxes}
\usage{
get.dominant(pr0)
}
\arguments{
\item{pr0}{matrix or data frame with recall values in the first column and precision
values in the second}
}
\value{
index of dominant points
}
\description{
The function calculates pareto-front of multiple peeling trajectories
}
\examples{
set.seed(1)
pr0 <- matrix(c(1,2,3,4,5,6,3,4,6,5), ncol = 2, byrow = 2)
pr <- matrix(sample(1:20, nrow(pr0)*2), ncol = 2)
pr
pr0
get.dominant(pr)
get.dominant(pr0)
}
\keyword{front}
\keyword{pareto}
