% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labs_box.R
\name{get.labs.box}
\alias{get.labs.box}
\title{Syntetic data generator}
\usage{
get.labs.box(box = NULL, n.points = 0, laths = TRUE, pts = 0,
  use.pts = FALSE, dgp, thr.hart3 = -1, thr.hart6sc = 1,
  thr.sobol = 0.7, thr.ishigami = 1, thr.morris = 20,
  thr.otlcircuit = 4.5, thr.borehole = 1000, thr.piston = 0.4,
  thr.wingweight = 250, thr.moon10hd = 0, thr.moon10hdc1 = 0,
  thr.moon10low = 1.5, thr.welchetal92 = 0, thr.linketal06dec = 0.15,
  thr.linketal06simple = 0.33, thr.linketal06sin = 0,
  thr.loepetal13 = 9, thr.morretal06 = -330, thr.oakoh04 = 10,
  thr.willetal06 = -1, thr.hart4 = -0.5, thr.ellipse = 0.8,
  thr.soblev99 = 2000)
}
\arguments{
\item{box}{matrix of real, specifying the box, from which the points should be sampled. If null, the unit box is used}

\item{n.points}{integer. Number of points to be sampled and labeled}

\item{laths}{logical. If true, use \code{randomLHS} sampling, else use random sampling}

\item{pts}{matrix of real, containing examples to be labeled. Only active if use.pts is true.}

\item{use.pts}{logical. If true, pre-spesified points \code{pts} are labeled, otherwize, new \code{n.points} 
points are sampled from the hyperbox \code{box}}

\item{dgp}{character. Specifies the DGP to be used. 
\itemize{
\item "ellipse" labels with 1 points inside 
an ellipse and with 0 otherwise.
\item Values from "1" to "10" for synthetic processess described in [1].
\item "borehole", "hart3", "hart4", "hart6sc", "ishigami", "linketal06dec", 
"linketal06simple", "linketal06sin", "loepetal13", "moon10hd", "moon10hdc1", "moon10low", "morretal06",
"oakoh04", "otlcircuit", "piston", "soblev99", "welchetal92", "willetal06", "wingweight" are the DGPs
specified in [2].
\item "sobol" and "morris" are the functions defined in R package 'sensitivity'
}}

\item{...}{parameters in the form \code{thr.x}, where \code{x} is one of possible \code{dgp} values.
Threshold value for the \code{x}-th DGP. Used to convert real function output into binary.}
}
\value{
list of two, containing a matrix of generated points and a vector of labels
}
\description{
The function samples points (if not provided) within a multi-dimensional hyperbox
and labels them according to pre-specified DGP
}
\examples{
d <- get.labs.box(n.points = 100, dgp = "1")
plot(d[[1]][, 1:2], col = d[[2]] + 1, pch = 20)

}
\references{
[1] Dalal, S. et al. 2013. Improving scenario discovery using orthogonal rotations. 
Environmental Modelling and Software. 48, (2013), 49-64.

[2] Surjanovic, S. & Bingham, D. (2013). Virtual Library of Simulation Experiments: Test Functions and Datasets.
\url{http://www.sfu.ca/~ssurjano}
}
