% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_norm_prim_wracc.R
\name{norm.prim.w}
\alias{norm.prim.w}
\title{PRIM returning the box maximizing WRAcc}
\usage{
norm.prim.w(dtrain, dtest = NULL, box, max.peels = 999, peel.alpha = 0.05)
}
\arguments{
\item{dtrain}{list, containing training data. The first element contains matrix/data frame of real attribute values.
the second element contains vector of labels 0/1.}

\item{dtest}{list, containing test data. Structured in the same way as \code{dtrain}}

\item{box}{matrix of real. Initial hyperbox, covering data}

\item{max.peels}{integer. Maximum length of the peeling trajectory (number of boxes)}

\item{peel.alpha}{real. The peeling parameter of PRIM from the interval (0,1)}
}
\value{
list.
\itemize{
\item \code{qtest} WRAcc value computed from \code{dtest}
\item \code{qtrain} WRAcc value computed from \code{dtrain}
\item \code{box} the hyperbox, with highest value of WRAcc on \code{dtrain}
}
}
\description{
The function applies PRIM to train data and evaluates its quality on test data
}
\examples{

dtrain <- dtest <- list()
dtest[[1]] <- dsgc_sym[1001:10000, 1:12]
dtest[[2]] <- dsgc_sym[1001:10000, 13]
dtrain[[1]] <- dsgc_sym[1:1000, 1:12]
dtrain[[2]] <- dsgc_sym[1:1000, 13]
box <- matrix(c(0.5,0.5,0.5,0.5,1,1,1,1,0.05,0.05,0.05,0.05,
5,5,5,5,4,4,4,4,1,1,1,1), nrow = 2, byrow = TRUE)

vals <- c(0.01, 0.03, 0.05, 0.07, 0.09, 0.11, 0.13, 0.15, 0.17, 0.19)
rtr <- rte <- numeric()
for(i in vals){
  tmp <- norm.prim.w(dtrain = dtrain, dtest = dtest, box = box, peel.alpha = i)
  rtr <- c(rtr, tmp[[2]])
  rte <- c(rte, tmp[[1]])
}

plot(vals, seq(min(c(rtr, rte)), max(c(rtr,rte)), along.with = vals), type = "n")
lines(vals, rtr, col = "red")
lines(vals, rte, col = "blue")

set.seed(1)
norm.prim.w(dtrain = dtrain, dtest = dtest, box = box, peel.alpha = 0.05)$qtest
norm.prim.w(dtrain = dtrain, dtest = dtest, box = box, peel.alpha = vals)$qtest
best.interval(dtrain = dtrain, dtest = dtest, box = box, depth = ncol(dtrain[[1]]), beam.size = 1)$qtest
}
\references{
Friedman, J.H. and Fisher, N.I. 1999. Bump hunting in high-dimensional data.
Statistics and Computing. 9, 2 (1999), 123-143.
}
\seealso{
\code{\link{norm.prim}},
\code{\link{best.interval}}
}
\keyword{models,}
\keyword{multivariate}
