% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_primsingle.R
\name{prim.single}
\alias{prim.single}
\title{One cut PRIM}
\usage{
prim.single(
  x,
  y,
  box,
  peel.alpha = 0.05,
  threshold = 1,
  pasting = FALSE,
  paste.alpha = 0.01,
  x.init,
  y.init
)
}
\arguments{
\item{x}{matrix of real. Data values for examples within the box}

\item{y}{vector of response values 0/1 for examples within the box}

\item{box}{matrix of real. Initial covering box}

\item{peel.alpha}{real. Peeling quantile tuning parameter}

\item{threshold}{real. Threshold tuning parameter. If average value of \code{y} exceeds it,
the box is not modyfied anymore.}

\item{pasting}{logical. If true pasting step is performed}

\item{paste.alpha}{real. Pasting quantile tuning parameter}

\item{x.init}{matrix of real. Data values for all examples}

\item{y.init}{vector of response values 0/1 for all examples}
}
\value{
list.
\itemize{
\item \code{meet.threshold} logical. If the threshold is reached
\item \code{res} a list of 3. Contains a new box obtained after one cut (if made),
\code{x} elements falling withinin it and resective vector of labels, \code{y}
\item \code{box.p} box after pasting step. If pasting is false, coincides with \code{res$box}
}
}
\description{
The function makes one cut following the PRIM algorithm, if the threshold is not met.
}
\examples{

box <- matrix(c(0.5, 0.5, 0.5, 0.5, 1, 1, 1, 1, 0.05, 0.05, 0.05, 0.05, 
5, 5, 5, 5, 4, 4, 4, 4, 1, 1, 1, 1), nrow = 2, byrow = TRUE)
x <- dsgc_sym[1:99, 1:12]
y <- dsgc_sym[1:99, 13]
res <- prim.single(x, y, box, pasting = TRUE, x.init = x, y.init = y)
res[[2]]$box

set.seed(1)
box <- matrix(c(0, 0, 1, 1), nrow = 2, byrow = TRUE)
x <- cbind(runif(1000), runif(1000))
y <- ifelse(x[, 1] > 0.98, 0, 1)
res <- prim.single(x, y, box, pasting = TRUE, x.init = x, y.init = y)
res[[2]]$box
res[[3]]
}
\keyword{models,}
\keyword{multivariate}
