% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_auc.R
\name{qual.auc}
\alias{qual.auc}
\title{AUC metric for PRIM}
\usage{
qual.auc(pr)
}
\arguments{
\item{pr}{matrix with coverage (recall) in the first column and 
density (precision) in the second column}
}
\value{
real. AUC value
}
\description{
The function calculates "area under the peeling curve" in "coverage-density" coordinates
}
\examples{

dtrain <- dtest <- list()
dtest[[1]] <- dsgc_sym[1001:10000, 1:12]
dtest[[2]] <- dsgc_sym[1001:10000, 13]
dtrain[[1]] <- dsgc_sym[1:500, 1:12]
dtrain[[2]] <- dsgc_sym[1:500, 13]
box <- matrix(c(0.5,0.5,0.5,0.5,1,1,1,1,0.05,0.05,0.05,0.05,
5,5,5,5,4,4,4,4,1,1,1,1), nrow = 2, byrow = TRUE)

set.seed(1)
res <- norm.prim(dtrain = dtrain, dtest = dtest, box = box)

# AUC on test data
qual.auc(res[[1]])

# AUC on train data
qual.auc(res[[2]])
}
