% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_interpretability.R
\name{qual.interpretability}
\alias{qual.interpretability}
\title{interpretability of PRIM result}
\usage{
qual.interpretability(res.box, box.init)
}
\arguments{
\item{res.box}{box to evaluate}

\item{box.init}{real. Matrix describing the initial box containing all data}
}
\value{
number of restricted dimensions in res.box as compared to box
}
\description{
The function calculates the quantity opposite to interpretability,
as the number of dimensions restricted. 
Higher value stands for lower interpretability
}
\examples{

box <- matrix(c(rep(0, 5), rep(1, 5)), ncol = 5, byrow = TRUE)
res.box <- matrix(c(0.40, 0.49, 0, 0.15, 0, 1, 1, 1, 0.93, 1), ncol = 5, byrow = TRUE)
box
res.box
qual.interpretability(res.box, box)
}
\references{
Bryant, B.P. and Lempert, R.J. 2010. Thinking inside the box: 
A participatory, computer-assisted approach to scenario discovery. 
Technological Forecasting and Social Change. 77, 1 (2010), 34-49.
}
