% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_wracc.R
\name{qual.wracc}
\alias{qual.wracc}
\title{WRAcc measure of the box}
\usage{
qual.wracc(d, box)
}
\arguments{
\item{d}{list containing the data. The first element contains matrix/data frame of real attribute values.
the second element contains vector of labels 0/1.}

\item{box}{matrix of real describing the box of interest}
}
\value{
WRAcc value
}
\description{
The function calculates the WRAcc of the box using a given data
}
\examples{

d <- list()
d[[1]] <- dsgc_sym[, 1:12]
d[[2]] <- dsgc_sym[, 13]
box <- matrix(c(1,1,1,1,1,1,1,1,0.05,0.05,0.05,0.05,
4,4,4,4,4,4,4,4,1,1,1,1), nrow = 2, byrow = TRUE)

qual.wracc(d, box)
}
