% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_select_depth.R
\name{select.depth}
\alias{select.depth}
\title{optimization of depth with WRAcc}
\usage{
select.depth(dtrain, box, depth, beam.size = 1, keep = 10)
}
\arguments{
\item{dtrain}{list, containing training data. The first element contains matrix/data frame of real attribute values.
the second element contains vector of labels 0/1.}

\item{box}{matrix of real. Initial hyperbox, covering data}

\item{depth}{a set of real. The peeling parameters of PRIM to test from the interval (0,1)}

\item{beam.size}{integer.The size of the beam in the beam search}

\item{keep}{integer. If greater than \code{beam.size}, specifies the maximum
number of boxes to be refined at each iteration in case all have equal WRAcc}
}
\value{
the optimal value of the peeling parameter from the set
}
\description{
The function chooses one depth value out of a given set using cross validation
and WRAcc as a quality measure
}
\examples{

dtrain <- list()
dtrain[[1]] <- dsgc_sym[1:1000, 1:12]
dtrain[[2]] <- dsgc_sym[1:1000, 13]
box <- matrix(c(0.5,0.5,0.5,0.5,1,1,1,1,0.05,0.05,0.05,0.05,
5,5,5,5,4,4,4,4,1,1,1,1), nrow = 2, byrow = TRUE)

select.depth(dtrain = dtrain, box = box, depth = c(4, 8, 12))
}
\seealso{
\code{\link{select.alpha.w}}
}
