% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_bagging_prim.R
\name{bagging.prim}
\alias{bagging.prim}
\title{PRIM with bagging (bumping)}
\usage{
bagging.prim(dtrain, dtest, box, features = "some", iter = 50,
  minpts = 20)
}
\arguments{
\item{dtrain}{list, containing training data. The first element contains matrix/data frame of real attribute values.
the second element contains vector of labels 0/1.}

\item{dtest}{list, containing test data. Structured in the same way as \code{dtrain}}

\item{box}{hyperbox, covering data}

\item{features}{character or integer. Refers to the number of features used in each iteration.
If integer, defines this number.
If "some", calculated as \code{ceiling(sqrt(ncol(dtrain[[1]])))},
if "all", all features are used. Default is "some"}

\item{iter}{integer. The number of iterations (sequences of boxes learned). Default is 50}

\item{minpts}{integer. Minimal number of points in the box for PRIM to continue peeling}
}
\value{
list.
\itemize{
\item \code{pr.test} matrix with coverage (recall) in the first column and 
density (precision) in the second column, evaluated on \code{dtest}
\item \code{pr.train} matrix with coverage (recall) in the first column and 
density (precision) in the second column, evaluated on \code{deval}
\item \code{boxes} list of matrices defining boxes constituting peeling trajectory
}
}
\description{
The function implements PRIM with bagging (bumping)
}
\examples{

dtrain <- dtest <- list()
dtest[[1]] <- dsgc_sym[1001:10000, 1:12]
dtest[[2]] <- dsgc_sym[1001:10000, 13]
dtrain[[1]] <- dsgc_sym[1:500, 1:12]
dtrain[[2]] <- dsgc_sym[1:500, 13]
box <- matrix(c(0.5,0.5,0.5,0.5,1,1,1,1,0.05,0.05,0.05,0.05,
5,5,5,5,4,4,4,4,1,1,1,1), nrow = 2, byrow = TRUE)

res <- norm.prim(dtrain = dtrain, dtest = dtest, box = box)
set.seed(1)
res.bag <- bagging.prim(dtrain = dtrain, dtest = dtest, box = box)
res.bag.all <- bagging.prim(dtrain = dtrain, dtest = dtest, box = box, features = 12)

plot(res[[1]], type = "l", xlim = c(0, 1), ylim = c(0.5, 1),
xlab = "recall", ylab = "precision")
lines(res[[2]], col = "red")
lines(res.bag[[1]], lty = 2)
lines(res.bag[[2]], lty = 2, col = "red")
lines(res.bag.all[[1]], lty = 3)
lines(res.bag.all[[2]], lty = 3, col = "red")
legend("bottomleft", legend = c("norm test", "norm train", "bagging test", 
"bagging train", "bagging all test", "bagging all train"),
lty = c(1, 1, 2, 2, 3, 3), col = c("black", "red", "black", "red", "black", "red"))

}
\references{
J. H. Kwakkel and S. C. Cunningham, "Improving scenario discovery by bagging random boxes," 
Technological Forecasting and Social Change, vol. 111, pp. 124-134, 2016.
}
\seealso{
\code{\link{norm.prim}},
\code{\link{rf.prim}}
}
\keyword{models,}
\keyword{multivariate}
