% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_select_param_bagging.R
\name{select.param}
\alias{select.param}
\title{optimization of peel.alpha and features with AUpC}
\usage{
select.param(dtrain, box, minpts = 20, peel.alpha, features, iter = 50)
}
\arguments{
\item{dtrain}{list, containing training data. The first element contains matrix/data frame of real attribute values.
the second element contains vector of labels 0/1.}

\item{box}{matrix of real. Initial hyperbox, covering data}

\item{minpts}{integer. Minimal number of points in the box for PRIM to continue peeling}

\item{peel.alpha}{real or a set of real. The peeling parameters of PRIM to test from the interval (0,1)}

\item{features}{character, integer, a set of characters or integers.
Refers to the number of features used in each iteration.
If integer, defines this number.
If "some", calculated as \code{ceiling(sqrt(ncol(dtrain[[1]])))},
if "all", all features are used. Default is "some"}

\item{iter}{integer. The number of iterations (sequences of boxes learned). Default is 30}
}
\value{
the optimal value of the peeling parameter from the set
}
\description{
The function chooses one peel.alpha value and one value of number of features
out of given sets for prim with bagging algorithm
using cross validation and AUpC as a quality measure
}
\examples{

dtrain <- list()
dtrain[[1]] <- dsgc_sym[1:400, 1:12]
dtrain[[2]] <- dsgc_sym[1:400, 13]
box <- matrix(c(0.5,0.5,0.5,0.5,1,1,1,1,0.05,0.05,0.05,0.05,
5,5,5,5,4,4,4,4,1,1,1,1), nrow = 2, byrow = TRUE)

res <- select.param(dtrain = dtrain, box = box, minpts = 20,
peel.alpha = c(0.17), iter = 20, features = c(1,4,12))
}
\seealso{
\code{\link{bagging.prim}},
}
